requirements.txt
google-generativeai 
langchain
langchain-google-genai 
chromadb 
pypdf
langchain-community 
langchain-core
python-dotenv